+++ src/lib.rs
@@ -1,5 +1,5 @@
 mod cli;
-mod connectors;
+pub mod connectors;
 pub mod monitor;
 mod net;
 mod providers;