+++ src/monitor/mod.rs
@@ -1,9 +1,8 @@
-use async_trait::async_trait;
-use clap::{Subcommand, ValueEnum};
-
 mod depth;
 mod order_book;
 
+use async_trait::async_trait;
+use clap::{Subcommand, ValueEnum};
 pub(crate) use depth::Depth;
 use depth::DepthConfig;
 
@@ -20,7 +19,7 @@ pub enum Provider {
     Other,
 }
 
-#[async_trait]
+#[async_trait(?Send)]
 pub(crate) trait Monitor {
     async fn monitor(&self);
 }